AI-Hand-Tracking/requirements.txt
opencv-python
customtkinter
Pillow
mediapipe
